requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.7
Flask-Migrate==4.0.4
python-dotenv==1.0.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1